Tidy up the cycle timer and add an animation test
Log only when the timer is actually armed, drop a redundant
timerclear, list A in the modifier comment and move its case next
to the other modifiers taking a number.  Revert the unused
options_table_entry widening left over from the negative
status-interval.  Add a regress test which watches an inner tmux
status line animate on its own.

+++ tmux.h
@@ -2441,8 +2441,8 @@ struct options_table_entry {
 	int			  scope;
 	int			  flags;
 
-	long long		  minimum;
-	long long		  maximum;
+	u_int			  minimum;
+	u_int			  maximum;
 	const char		**choices;
 
 	const char		 *default_str;
